Formatting

+++ demo/eth_verifier/src/VerifierIndex.sol
@@ -6,8 +6,8 @@ import {URS} from "./Commitment.sol";
 
 struct VerifierIndex {
     URS urs;
-    uint public_len;
-    uint domain_size;
-    uint max_poly_size;
+    uint256 public_len;
+    uint256 domain_size;
+    uint256 max_poly_size;
     BN254.G1Point blinding_commitment;
 }
